pyproject.toml
[project]
name = "daily"
version = "0.1.0"
description = "Post a daily NKJV quote to Slack using the xAI API"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
